requirements.txt
customtkinter>=5.2.2
Pillow>=10.0
cairosvg>=2.7
pyinstaller>=6.0
